chore: reorder the project's structure
Reorders the project's folders. Now, there is a static folder for the .js and .css files.
Also, there is a folder that contains the files to manage the views
