+++ crud/urls.py
@@ -1,6 +1,7 @@
 
 from django.urls import path
-from . import views
+from .views import views
+from .views import request_info_update
 
 urlpatterns = [
     path('', views.home, name='home'),
@@ -8,5 +9,5 @@ urlpatterns = [
     path('login/', views.logIn),
     path('postlog/', views.postLog),
     path('logout/', views.singout),
-    path('askInfoUpdate/', views.ask_info_update)
+    path('askInfoUpdate/', request_info_update.ask_info_update)
 ]
