+++ crud/views/views.py
@@ -6,7 +6,7 @@ from django.contrib.auth import login, authenticate, logout
 from django.db import IntegrityError
 from django.contrib.auth.decorators import login_required
 
-from .models import Office
+from ..models import Office
 
 import smtplib
 from email.mime.text import MIMEText
@@ -60,44 +60,3 @@ def postLog(request):
 def singout(request):
     logout(request)
     return redirect(home)
-
-
-@login_required
-def ask_info_update(request):
-    if request.method == 'POST':
-        selected_office_id = request.POST.get('offices')
-
-        try:
-            selected_office = Office.objects.get(id=selected_office_id)
-        except Office.DoesNotExist:
-            return HttpResponse('La oficina seleccionada no existe')
-
-        send_email(selected_office.email)
-
-        return redirect(home)
-
-    offices = Office.objects.all()
-
-    return render(request, 'ask_update_info.html', {
-        'offices' : offices
-    })
-
-
-def send_email(receiver_email):
-    subject = "Solicitud de actualización de información"
-    message = "Por favor, actualice la información, amiguito bello."
-
-    sender = "pi.seg.estudiantes@gmail.com"
-    password = "zhazuuahhicywuyg"
-
-    mail = MIMEText(message)
-    mail['From'] = sender
-    mail['To'] = receiver_email
-    mail['Subject'] = subject
-
-    with smtplib.SMTP("smtp.gmail.com", 587) as gmail_server:
-        gmail_server.starttls()
-        gmail_server.login(sender, password)
-
-        gmail_server.send_message(mail)
-
